Set "sideEffects" value to "true" in package.json.
The `sideEffects` value for this package should be `true`. Its current value of `false` causes the `import "ag-psd/initialize-canvas"` declaration to be dropped in bundlers that support tree-shaking, like Webpack and esbuild. Having `sideEffects: false` here is incorrect because that `import` statement in this package is explicitly used for its side-effects. This doesn't effect browser environments that use bundling, as they don't use that import, but it does effect bundled Node environments, like AWS Lambda deployments, which do use that import.